Suite modify/delete post + page profil + suppression profil

+++ backend/controllers/article.js
@@ -3,94 +3,75 @@ const User = require('../models/userSchema');
 const Comments = require('../models/commentSchema');
 
 const Op = require("sequelize").Op;
-
 const fs = require('fs');
 
 exports.createArticle = (req, res, next) => {
   const content = req.body.content;
-  const imageUrl = req.file.filename;
   const users_id = req.auth.userId;
   const articleObject = req.body;
-  console.log(imageUrl)
-    if(imageUrl == null) {
-      const article = { ...articleObject, users_id};
-      const post = new Article({
-        ...article})
-      post.save()
-        .then(() => res.status(201).json(article))
-        .catch(error => res.status(400).json({ error }));
-    }
-    else {
-        const article = {...articleObject, imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`, users_id};
-      const post = new Article({
-        ...article})
-        post.save(({ where: {id: req.params.id}}))
-          .then((article) => res.status(201).json(article))
-          .catch(error => res.status(400).json({ error }));
-    }
+  if((content == null || content == "") && (!req.file)){
+    return res.status(400).json(res.status);
+  }
+  else if(req.file) {
+    const article = {...articleObject, imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`, users_id};
+    const post = new Article({
+      ...article})
+    post.save(({ where: {id: req.params.id}}))
+      .then((article) => res.status(201).json(article))
+      .catch(error => res.status(400).json({ error }));
+  }
+  else {
+    const article = { ...articleObject, users_id};
+    const post = new Article({
+      ...article})
+    post.save()
+      .then(() => res.status(201).json(article))
+      .catch(error => res.status(400).json({ error }));
+  }
 }
     
-
 exports.getOneArticle = (req, res, next) => {
-    Article.findOne({ 
-      where:{
-        id: req.params.id 
+  Article.findOne({ 
+    where:{id: req.params.id},
+    include:[{model: User}]
+  })
+    .then((article) => res.status(200).json(article))
+    .catch((error) => res.status(404).json({ error }));
+}
+  
+exports.getAllArticles = (req, res, next) => {
+  where = {};
+  if(req.query.type == "text"){
+    where = {
+      where: {
+        content: {[Op.not]: null || ""}
       },
       include:[
-        {
-          model: User
-        }
+        {model: User},
+        {model: Comments}
       ]
-    })
-        .then((article) => res.status(200).json(article))
-        .catch((error) => res.status(404).json({ error }));
-    };
-  
-  exports.getAllArticles = (req, res, next) => {
-    where = {};
-    if(req.query.type == "text"){
-      where = {
-        where: {
-          content: {
-            [Op.not]: null
-          }
-        },
-        include:[
-          {
-            model: User
-          },
-          {
-            model: Comments
-          }
-        ]
-      }
     }
-    if(req.query.type == "image"){
-      where = {
-        where: {
-          imageUrl: {
-            [Op.not]: ""
-          }
-        },
-        include:[
-          {
-            model: User
-          },
-          {
-            model: Comments
-          }
-        ]
-      }
+  }
+  if(req.query.type == "image"){
+    where = {
+      where: {
+        imageUrl: {[Op.not]: ""}
+      },
+      include:[
+        {model: User},
+        {model: Comments}
+      ]
     }
-      Article.findAll(where)
-          .then((articles) => res.status(200).json(articles))
-          .catch((error) => res.status(400).json({ error }));
-      };
+  }
+  Article.findAll(where)
+    .then((articles) => res.status(200).json(articles))
+    .catch((error) => res.status(400).json({ error }));
+}
 
 exports.modifyArticle = (req, res, next) => {
   const articleObject = req.file ?
   {
-    ...JSON.parse(req.body.article),
+    ...req.body,
     imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
   } : { ...req.body };
   Article.findOne({ where: {id: req.params.id}})
@@ -103,10 +84,22 @@ exports.modifyArticle = (req, res, next) => {
           error: new Error('Unauthorized request !')
         });
       }
+      if(article.imageUrl != null){
+        console.log(req.file)
+        const filename = article.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Article.update({ ...articleObject, id:  req.params.id},{where: {id: req.params.id}})
+            .then(() => res.status(200).json({ message: 'Article modifié !'}))
+            .catch(error => res.status(400).json({ error }));
+        })
+      }
+      else {
+        console.log(req.body.content)
         Article.update({ ...articleObject, id:  req.params.id},{where: {id: req.params.id}})
           .then(() => res.status(200).json({ message: 'Article modifié !'}))
           .catch(error => res.status(400).json({ error }));
-      })
+      }
+    })
     .catch(error => res.status(500).json({ error }));
   };
 
@@ -121,9 +114,19 @@ exports.deleteArticle = (req, res, next) => {
         error: new Error('Unauthorized request !')
       })
     }
-    Article.destroy({ where: {id: req.params.id} })
-    .then((article) => res.status(204).json(article))
-      .catch(error => res.status(400).json({ error }));
+    if(article.imageUrl != null){
+    const filename = article.imageUrl.split('/images/')[1];
+    fs.unlink(`images/${filename}`, () => {
+      Article.destroy({ where: {id: req.params.id} })
+        .then((article) => res.status(204).json(article))
+        .catch(error => res.status(400).json({ error }));
+      })
+    }
+    else {
+      Article.destroy({ where: {id: req.params.id} })
+        .then((article) => res.status(204).json(article))
+        .catch(error => res.status(400).json({ error }));
+      }
     })
   .catch(error => res.status(500).json({ error }));
 };
